+++ spark-bigquery-dsv1/spark-bigquery-with-dependencies_2.12/src/test_2.12/java/com/google/cloud/spark/bigquery/acceptance/Scala212ReadSheakspeareDataprocServerlessAcceptanceTest.java
@@ -18,7 +18,7 @@ package com.google.cloud.spark.bigquery.acceptance;
 public class Scala212ReadSheakspeareDataprocServerlessAcceptanceTest extends ReadSheakspeareDataprocServerlessAcceptanceTestBase {
 
     public Scala212ReadSheakspeareDataprocServerlessAcceptanceTest() {
-        super("spark-bigquery");
+        super("spark-bigquery", "1.0");
     }
 
     // tests from superclass