+++ spark-bigquery-pushdown/spark-2.4-bigquery-pushdown_2.11/src/test/scala/com/google/cloud/spark/bigquery/pushdowns/Spark24ExpressionConverterSuite.scala
@@ -1,10 +1,25 @@
+/*
+ * Copyright 2022 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.google.cloud.spark.bigquery.pushdowns
 
 import com.google.cloud.spark.bigquery.direct.DirectBigQueryRelation
-import org.apache.spark.sql.catalyst.expressions.{AttributeReference, CheckOverflow, ExprId, Literal, ScalarSubquery, UnaryMinus}
+import org.apache.spark.sql.catalyst.expressions.{AttributeReference, CheckOverflow, ExprId, Like, Literal, ScalarSubquery, UnaryMinus}
 import org.apache.spark.sql.catalyst.plans.logical.Aggregate
 import org.apache.spark.sql.execution.datasources.LogicalRelation
-import org.apache.spark.sql.types.{DecimalType, LongType, StructType}
+import org.apache.spark.sql.types.{DecimalType, LongType, StringType, StructType, TimestampType}
 import org.mockito.Mockito.when
 import org.mockito.{Mock, MockitoAnnotations}
 import org.scalatest.BeforeAndAfter
@@ -47,5 +62,20 @@ class Spark24ExpressionConverterSuite extends AnyFunSuite with BeforeAndAfter {
     val bigQuerySQLStatement = expressionConverter.convertMathematicalExpressions(checkOverflowExpression, fields)
     assert(bigQuerySQLStatement.isDefined)
     assert(bigQuerySQLStatement.get.toString == "CAST ( 233.45 AS BIGDECIMAL )")
+  }
+
+  test("convertBasicExpressions with Timestamp literal") {
+    // Internally, a timestamp is stored as the number of microseconds from the epoch of 1970-01-01T00
+    val bigQuerySQLStatement = expressionConverter.convertBasicExpressions(Literal(1230219000000000L, TimestampType), fields)
+    assert(bigQuerySQLStatement.isDefined)
+    assert(bigQuerySQLStatement.get.toString == "TIMESTAMP_MICROS( 1230219000000000 )")
+  }
+
+  test("convertStringExpressions with Like") {
+    val bigQuerySQLStatement = expressionConverter.convertStringExpressions(
+      Like(AttributeReference.apply("SchoolID", StringType)(ExprId.apply(1)),
+        Literal("%aug%urs%")), fields)
+    assert(bigQuerySQLStatement.isDefined)
+    assert(bigQuerySQLStatement.get.toString == "SUBQUERY_2.SCHOOLID LIKE '%aug%urs%'")
   }
 }