+++ spark-bigquery-pushdown/spark-3.2-bigquery-pushdown_2.13/src/main/scala/com/google/cloud/spark/bigquery/pushdowns/Spark32ExpressionConverter.scala
@@ -17,7 +17,7 @@
 package com.google.cloud.spark.bigquery.pushdowns
 
 import com.google.cloud.spark.bigquery.pushdowns.SparkBigQueryPushdownUtil.blockStatement
-import org.apache.spark.sql.catalyst.expressions.{Attribute, Cast, CheckOverflow, Expression, ScalarSubquery, UnaryMinus}
+import org.apache.spark.sql.catalyst.expressions.{Attribute, Cast, CheckOverflow, Expression, Like, ScalarSubquery, UnaryMinus}
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 
 /**
@@ -56,6 +56,13 @@ class Spark32ExpressionConverter(expressionFactory: SparkExpressionFactory, spar
     expression match {
       case Cast(child, dataType, _, ansiEnabled) if !ansiEnabled =>
         performCastExpressionConversion(child, fields, dataType)
+    }
+  }
+
+  override def convertLikeExpression(expression: Expression, fields: Seq[Attribute]): BigQuerySQLStatement = {
+    expression match {
+      case Like(left, right, _) =>
+        convertStatement(left, fields) + "LIKE" + convertStatement(right, fields)
     }
   }
 }